fix: address analytics refactor review comments
